+++ libs/ui/svelte/components/Icon/MissingIcon.svelte
@@ -73,6 +73,7 @@
   }
 
   :global(.slu-icon-outer img) {
+    width: 100%;
     height: 100%;
     object-fit: contain;
   }