+++ src/background/modules/monitors/domain.rs
@@ -1,19 +1,15 @@
 use seelen_core::system_state::PhysicalMonitor;
 
-use crate::{
-    error::AppError,
-    windows_api::{monitor::Monitor, WindowsApi},
-};
+use crate::{error::AppError, windows_api::monitor::Monitor};
 
 impl TryFrom<Monitor> for PhysicalMonitor {
     type Error = AppError;
     fn try_from(m: Monitor) -> Result<Self, Self::Error> {
-        let dpi = WindowsApi::get_monitor_scale_factor(m.handle())?;
         Ok(Self {
             id: m.stable_id()?.into(),
             name: m.name()?,
             rect: m.rect()?,
-            dpi,
+            scale_factor: m.scale_factor()?,
             is_primary: m.is_primary(),
         })
     }