+++ src/background/virtual_desktops/cli.rs
@@ -1,7 +1,10 @@
-use seelen_core::system_state::MonitorId;
+use seelen_core::{state::WidgetTriggerPayload, system_state::MonitorId};
 use serde::{Deserialize, Serialize};
 
-use crate::{error::Result, virtual_desktops::SluWorkspacesManager2, windows_api::window::Window};
+use crate::{
+    error::Result, virtual_desktops::SluWorkspacesManager2, widgets::trigger_widget,
+    windows_api::window::Window,
+};
 
 /// Manage the Seelen Window Manager.
 #[derive(Debug, Serialize, Deserialize, clap::Args)]
@@ -36,6 +39,8 @@ pub enum VdCommand {
     CreateNewWorkspace,
     /// Destroy the current workspace (will do nothing if there's only one workspace)
     DestroyCurrentWorkspace,
+    /// Toggle the workspace view
+    ToggleWorkspacesView,
 }
 
 impl VirtualDesktopCli {
@@ -112,6 +117,11 @@ impl VdCommand {
                     .get(&monitor_id, |monitor| monitor.active_workspace_id().clone())
                     .ok_or("Monitor not found")?;
                 vd.destroy_desktop(&monitor_id, &workspace_id)?;
+            }
+            VdCommand::ToggleWorkspacesView => {
+                trigger_widget(WidgetTriggerPayload::new(
+                    "@seelen/workspaces-viewer".into(),
+                ))?;
             }
         }
         Ok(())