+++ src/ui/svelte/workspaces-viewer/index.ts
@@ -8,9 +8,19 @@ import "@shared/styles/colors.css";
 const widget = Widget.getCurrent();
 await widget.init();
 
+// play with zoom level to reset device pixel ratio to 1:1
+await widget.webview.setZoom(1 / (await widget.webview.scaleFactor()));
+widget.webview.onScaleChanged(({ payload }) => {
+  widget.webview.setZoom(1 / payload.scaleFactor);
+});
+
 widget.webview.setResizable(false);
-widget.onTrigger(() => {
-  widget.webview.show();
+widget.onTrigger(async () => {
+  if (await widget.webview.isVisible()) {
+    widget.webview.hide();
+  } else {
+    widget.webview.show();
+  }
 });
 
 const root = document.getElementById("root")!;